refactor: change one tap button to one tap opener
